Removed ClearingMechanismActor.
